qa/Autotests/API/requirements.txt
psycopg2_binary==2.9.9
Requests==2.32.3
